requirements.txt
streamlit
google-generativeai
python-dotenv
openai-whisper
pytubefix
moviepy
librosa